Move organization mounting into `UserAuth`
Might possibly be better as a separate module/hook but since we want
to use `personal` as the slug for personal orgs, it has to be mounted with
current user available. Just seems cleaner to do it together rather than
having it pull `current_user` from the socket, which would mean one must
remember to 1) set the `mount_current_user` hook and 2) make sure it runs
before mounting the current org.